Update BedethequeScraper2.py
Formatting articles also on when serie name contains ":" and not only "-".
For example:
"The Authority ꞉ Les années Stormwatch" --> "Authority (The) ꞉ Les années Stormwatch"

When user choice dialog is proposed in case several serie/album have similar name, I have added a filter on top of this dialog. Only Serie/Album containing the characters on this filter will be listed. It allow to quickly find the correct serie when there is a lot of choices proposed.

Change all tabs into space.